requirements.txt
openai
tiktoken
python-dotenv~=1.0.0
fastapi~=0.110.2
uvicorn==0.24.0.post1
pydantic
presidio-analyzer==2.2.351
presidio_anonymizer==2.2.351
llama-index
llama-index-vector-stores-faiss
llama-index-tools-duckduckgo
llama-index-agent-openai
pypdf
requests
pandas
tqdm
pymongo
streamlit
faiss-cpu
langchain-community==0.3.27
ddgs==9.4.3
duckduckgo-search==8.1.1